Scraper/requirements.txt
plotly
pandas
redis
dash
celery
google-cloud-language
